+++ nbtdt/NBDownLoadManagerViewController.m
@@ -115,7 +115,7 @@
     }
     
     
-    int index=[_downlist.allKeys indexOfObject:[downItem.url description]];
+    NSInteger index=[_downlist.allKeys indexOfObject:[downItem.url description]];
     DowningCell *cell=(DowningCell *)[self.table cellForRowAtIndexPath:[NSIndexPath indexPathForRow:index inSection:0]];
     [self updateCell:cell withDownItem:downItem];
 }
