+++ nbtdt/StringUtil.m
@@ -26,7 +26,7 @@ static const char encodingTable[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopq
         return @"";
 
     const char *source = [str UTF8String];
-    int strlength  = strlen(source);
+    NSInteger strlength  = strlen(source);
     
     char *characters = malloc(((strlength + 2) / 3) * 4);
     if (characters == NULL)
@@ -57,8 +57,8 @@ static const char encodingTable[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopq
 {
 	NSMutableString* s = [NSMutableString string];
 	
-	int start = 0;
-	int len = [self length];
+	NSInteger start = 0;
+	NSInteger len = [self length];
 	NSCharacterSet* chs = [NSCharacterSet characterSetWithCharactersInString:@"<>&\""];
 	
 	while (start < len) {
